+++ sunpy/coordinates/__init__.py
@@ -3,6 +3,6 @@ from __future__ import absolute_import, division
 from .frames import *
 from .offset_frame import *
 from .transformations import *
-from .utilities import *
+from .ephemeris import *
 
 from .wcs_utils import *
