+++ sunpy/coordinates/ephemeris.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+"""
+Ephemeris calculations using SunPy coordinate frames
+
+"""
 from __future__ import absolute_import, division
 import datetime
 import warnings
